% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr.R
\docType{data}
\name{cr}
\alias{cr}
\title{Complete Realization data}
\format{
A data frame with 250.000 rows and 3 variables:
\describe{
\item{x-coordinates}
\item{y-coordinates}
\item{variable}
}
}
\usage{
cr
}
\description{
Contains coordinates and one variable.
}
\keyword{datasets}
