% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_cr.R
\name{generate_cr}
\alias{generate_cr}
\title{Generate a complete realization}
\usage{
generate_cr(true.nugget = 60, true.sigma_sq = 40, true.phi = 200, nmax = 100)
}
\arguments{
\item{true.nugget}{True nugget parameter}

\item{true.sigma_sq}{True partial sill parameter}

\item{true.phi}{True range parameter}

\item{nmax}{Maximum number of pairwise correlations?}
}
\description{
Generate a complete realization
}
