% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_scenario.R
\name{load_scenario}
\alias{load_scenario}
\title{Load single scenario (without summarizing)}
\usage{
load_scenario(
  wd = "D:/Sciebo/EgoCorSim_mit_Ole",
  N = 500,
  dens = 1,
  max.dist = 457,
  n_runs = 1:10,
  filter = rep(1000, 3)
)
}
\arguments{
\item{wd}{working directory were ALL the simulation files lie}

\item{N}{Number of observations (500, 1000, 2000)}

\item{dens}{Sampling density (1, 4, 9)}

\item{max.dist}{Maximum distance (457, 624, 832)}

\item{n_runs}{Number of seperate files for each scenario}

\item{filter}{Vector with upper non-convergence threshold for kicking out models}
}
\value{
Returns a list containing the n's and the result
}
\description{
Load single scenario (without summarizing)
}
