% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_conv_table.R
\name{make_convtable}
\alias{make_convtable}
\title{convergence rate table}
\usage{
make_convtable(n_runs = 1:10, summary = T, kick_equal0 = F)
}
\arguments{
\item{n_runs}{Number of seperate files for each scenario}

\item{wd}{working directory were ALL the simulation files lie}

\item{N}{Number of observations (500, 1000, 2000)}

\item{dens}{Sampling density (1, 4, 9)}

\item{max.dist}{Maximum distance (457, 624, 832)}

\item{filter}{Vector with upper non-convergence threshold for kicking out models}
}
\value{
Returns a list containing the n's and the result
}
\description{
convergence rate table
}
