% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_table.R
\name{make_table}
\alias{make_table}
\title{Creating result table for simulation study fast}
\usage{
make_table(wd, param, method = "check", n_runs)
}
\arguments{
\item{wd}{working directory were ALL the simulation files lie}

\item{param}{Selects the desired parameter (1 = nugget, 2 = partial_sill, 3 = shape)}

\item{method}{one of c("check", "quantile", "brisc")}

\item{n_runs}{Number of seperate files for each scenario (this needs to fit to the largest integer at the end of the file names)}
}
\value{
A list containing the tables
}
\description{
Creating result table for simulation study fast
}
