% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_scenario.R
\name{one_scenario}
\alias{one_scenario}
\title{Title}
\usage{
one_scenario(
  input,
  cr,
  nbins,
  B = 1000,
  threshold.factor = c(1.1, 1.2, 1.5, 2, 2.5, 3),
  qu = seq(from = 0.75, to = 1, by = 0.05),
  n_sim
)
}
\arguments{
\item{input}{Input vector: Consisting of N (Number of samples), nr_divisions (for spatial_sampling function) and max.dist}

\item{cr}{Complete realization}

\item{nbins}{Number of bins for empirical semivariogram estimation}

\item{B}{Number of bootstrap samples}

\item{threshold.factor}{threshold vector}

\item{qu}{quantile vector}

\item{n_sim}{number of resamples to draw}
}
\value{
Saves an RData file to the working directory
}
\description{
Title
}
