% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par_unc_q.R
\name{par_unc_q}
\alias{par_unc_q}
\title{Semi-variogram parameter uncertainty - Quantile method}
\usage{
par_unc_q(
  sample.geo,
  max.dist,
  nbins = 10,
  B = 1000,
  qu = seq(from = 0.75, to = 1, by = 0.05)
)
}
\arguments{
\item{qu}{}
}
\description{
Semi-variogram parameter uncertainty - Quantile method
}
