% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par_uncertainty_brisc.R
\name{par_uncertainty_brisc}
\alias{par_uncertainty_brisc}
\title{Semi-variogram parameter uncertainty - BRISC}
\usage{
par_uncertainty_brisc(
  sample,
  B = 1000,
  h = 1,
  n_omp = 1,
  init = "Initial",
  verbose = TRUE,
  nugget_status = 1
)
}
\arguments{
\item{B}{Number of bootstrap samples}

\item{nugget_status}{}

\item{BRISC_Out}{An Object created by BRISC_estimation}
}
\description{
Semi-variogram parameter uncertainty - BRISC
}
