% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par_uncertainty_q.R
\name{par_uncertainty_q}
\alias{par_uncertainty_q}
\title{Semi-variogram parameter uncertainty - Quantile method}
\usage{
par_uncertainty_q(sample.geo, max.dist, nbins = 10, B = 1000, alpha = 0.75)
}
\arguments{
\item{qu}{}
}
\description{
Semi-variogram parameter uncertainty - Quantile method
}
