% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_sampling.R
\name{spatial_sampling}
\alias{spatial_sampling}
\title{Function for spatial sampling from complete realization}
\usage{
spatial_sampling(cr, N, nr_divisions)
}
\arguments{
\item{cr}{complete realization}

\item{N}{Number of samples}

\item{nr_divisions}{Number of subdivisions: Choose between 1, 4, 9}
}
\value{
Returns a sample of length N
}
\description{
Function for spatial sampling from complete realization
}
