% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_q.R
\name{summarise_q}
\alias{summarise_q}
\title{Compute summary statistics for the simulation of the quantile based method}
\usage{
summarise_q(
  wd = "/Users/jan-ole/R/Boot",
  N = NULL,
  density = NULL,
  max.dist = NULL,
  qu = NULL,
  param,
  n_runs = 10,
  filter = c(1000, 1000, 1000),
  kick_equal0 = F
)
}
\arguments{
\item{wd}{working directory were ALL the simulation files lie}

\item{N}{Number of observations (500, 1000, 2000)}

\item{density}{Sampling density (1, 4, 9)}

\item{max.dist}{Maximum distance (457, 624, 832)}

\item{qu}{Quantile (here only scalar possible)}

\item{param}{Selects the desired parameter (1 = nugget, 2 = partial_sill, 3 = shape)}

\item{n_runs}{Number of seperate files for each scenario}

\item{filter}{Vector for kicking out models}
}
\value{
Returns a list containing the n's and the result
}
\description{
Compute summary statistics for the simulation of the quantile based method
}
