% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par_unc_brisc.R
\name{par_uncertainty_brisc}
\alias{par_uncertainty_brisc}
\title{Semi-variogram parameter uncertainty - BRISC}
\usage{
par_uncertainty_brisc(
  BRISC_Out,
  B = 1000,
  h = 1,
  n_omp = 1,
  init = "Initial",
  verbose = TRUE,
  nugget_status = 1
)
}
\arguments{
\item{nugget_status}{}
}
\description{
Semi-variogram parameter uncertainty - BRISC
}
