% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par_uncertainty_q.R
\name{par_uncertainty_q}
\alias{par_uncertainty_q}
\title{Semi-variogram parameter uncertainty - Quantile method}
\usage{
par_uncertainty_q(
  sample.geo,
  max.dist,
  nbins = 10,
  B = 1000,
  qu = seq(from = 0.75, to = 1, by = 0.05)
)
}
\arguments{
\item{sample.geo}{A spatial sample}

\item{max.dist}{Maximum distance included in the estimation}

\item{B}{Number of bootstrap samples}

\item{qu}{Sequence of quantiles}
}
\description{
Semi-variogram parameter uncertainty - Quantile method
}
