% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_check_hist.R
\name{summarise_check_hist}
\alias{summarise_check_hist}
\title{Compute summary statistics for the simulation of the check based method}
\usage{
summarise_check_hist(
  wd = "/Users/jan-ole/R/Boot",
  thr = 1.1,
  path = "/Users/jan-ole/R/Boot",
  n_runs = 10,
  filter = c(1000, 1000, 1000),
  kick_equal0 = F
)
}
\arguments{
\item{wd}{working directory were ALL the simulation files lie}

\item{thr}{Threshold factor (here only scalar possible)}

\item{path}{Path the the directory where plot is saved}

\item{n_runs}{Number of seperate files for each scenario}

\item{filter}{Vector for kicking out models}
}
\value{
Returns a list containing the n's and the result
}
\description{
Compute summary statistics for the simulation of the check based method
}
